Code Review: Fix 12 bugs found in full codebase review (round 1)
- Fix MQTT payload used without null-termination (mosquitto_comms.c)
- Fix audio thread search skipping thread 0 and all odd threads
- Fix audio threads dying permanently on bad filenames
- Replace strcpy with safe_strncpy in audio thread (buffer overflow)
- Fix float->int truncation in transition scale save/restore
- Fix font cache null-terminating wrong linked list node
- Fix "Invalid HD" typo in hud_manager log message
- Remove duplicate init_video_out_data() call
- Fix log_command strncpy missing null-termination
- Redact stream key from GStreamer pipeline log output
- Add overflow check and 16MB limit to curl download callback
- Replace sprintf with snprintf in detectnet.cpp
- Add safe_strncpy utility header (ported from DAWN common lib)

+++ src/rendering/element_renderer.c
@@ -1974,7 +1974,7 @@ void render_element_with_scale(element *curr_element, float scale, float alpha) 
    curr_element->in_transition = 1;
    curr_element->transition_alpha = alpha;
 
-   int original_scale = curr_element->scale;
+   float original_scale = curr_element->scale;
    curr_element->scale = scale;
 
    /* Save original dimensions and position */
